// lib/types.ts
export type Agent = {
  name: string;
  phone: string;
  address: string;
};

export type Region = {
  name: string;
  agents: { [key: string]: Agent };
};
// ...
export type AppSettings = {
    banners: string[];
    supportNumbers: {
        libyan: string;
        egyptian: string;
    };
    regions: { [key: string]: Region };
};